Added ettermiddagen logo and new links in header
