+++ src/components/countdown.js
@@ -5,7 +5,6 @@ import styled from 'styled-components'
 const Countdown = () => {
 
     const [date, setDate] = useState("")
-    const [count, setCount] = useState(0)
 
     useEffect(() => {
             const interval = setInterval(() => {
@@ -33,10 +32,6 @@ const Countdown = () => {
 
             <div>
                 <Number>{date}</Number>
-                <p>You clicked {count} times</p>
-                <button onClick={() => setCount(count+1)}>
-                    Click me
-                </button>
             </div>
         </div>
     )
